services/news/pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.6.1",
    "quixstreams>=3.5.0",
    "rarfile>=4.2",
    "requests>=2.32.3",
    "wget>=3.2",
]
